Refactor after review
