+++ ui/client/components/graph/node-modal/editors/Validators.ts
@@ -11,18 +11,13 @@ export type Validator = {
   description: string
 }
 
-export const validators: Record<validatorType, Function> = {
-  [validatorType.NOT_EMPTY_VALIDATOR]: () => notEmptyValidator,
-  [validatorType.ERROR_VALIDATOR]: (errors, fieldName) => errorValidator(errors, fieldName),
-}
+const error = (errors, fieldName) => errors && errors.find(error => error.fieldName === fieldName || error.fieldName === `$${fieldName}`)
 
 export const errorValidator = (errors, fieldName) => ({
   isValid: () => !error(errors, fieldName),
   message: error(errors, fieldName)?.message,
   description: error(errors, fieldName)?.description,
 })
-
-const error = (errors, fieldName) => errors && errors.find(error => error.fieldName === fieldName || error.fieldName === `$${fieldName}`)
 
 export const notEmptyValidator: Validator = {
   isValid: value => !_.isEmpty(value),
@@ -32,4 +27,9 @@ export const notEmptyValidator: Validator = {
 
 export function allValid(validators: Array<Validator>, values: Array<any>): boolean {
   return validators.every(validator => validator.isValid(...values))
+}
+
+export const validators: Record<validatorType, Function> = {
+  [validatorType.NOT_EMPTY_VALIDATOR]: () => notEmptyValidator,
+  [validatorType.ERROR_VALIDATOR]: (errors, fieldName) => errorValidator(errors, fieldName),
 }
